Fix Bandit MPS and Triton frozen build: float64 buffers, inspect fallback stub
- Bandit: .float() on audio window buffers and freq_weights (MPS no float64)
- Workers: inspect fallback stub def __frozen_noop__(x) for Triton compatibility
  (varargs *a,**kw rejected by Triton compiled functions)
- build_runtime_mac_mps: add -y, freeze, --skip-env

+++ models/bandit/core/model/bsrnn/maskestim.py
@@ -251,7 +251,7 @@ class OverlappingMaskEstimationModule(MaskEstimationModuleBase):
 
         if freq_weights is not None:
             for i, fw in enumerate(freq_weights):
-                self.register_buffer(f"freq_weights/{i}", fw)
+                self.register_buffer(f"freq_weights/{i}", fw.float())
 
                 self.use_freq_weights = use_freq_weights
         else:
